backend/requirements.txt
uvicorn~=0.27.0
fastapi~=0.109.0
httpx~=0.26.0
python-dotenv~=1.0.1
sqlalchemy~=2.0.25
pymysql~=1.1.0
python-multipart~=0.0.9
pytz~=2024.1
pytest~=8.0.2
gunicorn~=21.2.0